Add an EnigmaMappingParsingTest
