+++ intellij-plugin/src/test/java/org/quiltmc/intellij/enigma/language/EnigmaMappingParsingTest.java
@@ -1,0 +1,9 @@
+package org.quiltmc.intellij.enigma.language;
+
+import org.quiltmc.intellij.test.language.AbstractParsingTestCase;
+
+public class EnigmaMappingParsingTest extends AbstractParsingTestCase {
+	public EnigmaMappingParsingTest() {
+		super("enigma", "mapping", new EnigmaMappingParserDefinition());
+	}
+}
