+++ intellij-plugin/src/test/testData/enigma/ParsingTestData.txt
@@ -1,0 +1,86 @@
+EnigmaMappingFile: ParsingTestData.mapping(0,214)
+  EnigmaMappingClazzImpl(CLAZZ)(0,214)
+    PsiElement(CLASS)('CLASS')(0,5)
+    PsiElement(NEW_LINE)('\n')(5,6)
+    PsiElement(TAB)('\t')(6,7)
+    EnigmaMappingCommentImpl(COMMENT)(7,70)
+      PsiElement(COMMENT)('COMMENT')(7,14)
+      PsiWhiteSpace(' ')(14,15)
+      PsiElement(ANY)('lorem')(15,20)
+      PsiWhiteSpace(' ')(20,21)
+      PsiElement(ANY)('ipsum')(21,26)
+      PsiWhiteSpace(' ')(26,27)
+      PsiElement(ANY)('dolor')(27,32)
+      PsiWhiteSpace(' ')(32,33)
+      PsiElement(ANY)('sit')(33,36)
+      PsiWhiteSpace(' ')(36,37)
+      PsiElement(ANY)('amet')(37,41)
+      PsiWhiteSpace(' ')(41,42)
+      PsiElement(ANY)('consectetur')(42,53)
+      PsiWhiteSpace(' ')(53,54)
+      PsiElement(ANY)('adipiscing')(54,64)
+      PsiWhiteSpace(' ')(64,65)
+      PsiElement(ANY)('elit,')(65,70)
+    PsiElement(NEW_LINE)('\n')(70,71)
+    PsiElement(TAB)('\t')(71,72)
+    EnigmaMappingCommentImpl(COMMENT)(72,146)
+      PsiElement(COMMENT)('COMMENT')(72,79)
+      PsiWhiteSpace(' ')(79,80)
+      PsiElement(ANY)('sed')(80,83)
+      PsiWhiteSpace(' ')(83,84)
+      PsiElement(ANY)('do')(84,86)
+      PsiWhiteSpace(' ')(86,87)
+      PsiElement(ANY)('eiusmod')(87,94)
+      PsiWhiteSpace(' ')(94,95)
+      PsiElement(ANY)('tempor')(95,101)
+      PsiWhiteSpace(' ')(101,102)
+      PsiElement(ANY)('incididunt')(102,112)
+      PsiWhiteSpace(' ')(112,113)
+      PsiElement(ANY)('ut')(113,115)
+      PsiWhiteSpace(' ')(115,116)
+      PsiElement(ANY)('labore')(116,122)
+      PsiWhiteSpace(' ')(122,123)
+      PsiElement(ANY)('et')(123,125)
+      PsiWhiteSpace(' ')(125,126)
+      PsiElement(ANY)('dolore')(126,132)
+      PsiWhiteSpace(' ')(132,133)
+      PsiElement(ANY)('magna')(133,138)
+      PsiWhiteSpace(' ')(138,139)
+      PsiElement(ANY)('aliqua.')(139,146)
+    PsiElement(NEW_LINE)('\n')(146,147)
+    PsiElement(TAB)('\t')(147,148)
+    EnigmaMappingMethodImpl(METHOD)(148,154)
+      PsiElement(METHOD)('METHOD')(148,154)
+    PsiElement(NEW_LINE)('\n')(154,155)
+    PsiElement(TAB)('\t')(155,156)
+    EnigmaMappingClazzImpl(CLAZZ)(156,207)
+      PsiElement(CLASS)('CLASS')(156,161)
+      PsiElement(NEW_LINE)('\n')(161,162)
+      PsiElement(TAB)('\t')(162,163)
+      PsiElement(TAB)('\t')(163,164)
+      EnigmaMappingCommentImpl(COMMENT)(164,175)
+        PsiElement(COMMENT)('COMMENT')(164,171)
+        PsiWhiteSpace(' ')(171,172)
+        PsiElement(ANY)('foo')(172,175)
+      PsiElement(NEW_LINE)('\n')(175,176)
+      PsiElement(TAB)('\t')(176,177)
+      PsiElement(TAB)('\t')(177,178)
+      EnigmaMappingMethodImpl(METHOD)(178,199)
+        PsiElement(METHOD)('METHOD')(178,184)
+        PsiElement(NEW_LINE)('\n')(184,185)
+        PsiElement(TAB)('\t')(185,186)
+        PsiElement(TAB)('\t')(186,187)
+        PsiElement(TAB)('\t')(187,188)
+        EnigmaMappingCommentImpl(COMMENT)(188,199)
+          PsiElement(COMMENT)('COMMENT')(188,195)
+          PsiWhiteSpace(' ')(195,196)
+          PsiElement(ANY)('bar')(196,199)
+      PsiElement(NEW_LINE)('\n')(199,200)
+      PsiElement(TAB)('\t')(200,201)
+      PsiElement(TAB)('\t')(201,202)
+      EnigmaMappingFieldImpl(FIELD)(202,207)
+        PsiElement(FIELD)('FIELD')(202,207)
+    PsiElement(NEW_LINE)('\n')(207,208)
+    PsiElement(TAB)('\t')(208,209)
+    EnigmaMappingFieldImpl(FIELD)(209,214)
+      PsiElement(FIELD)('FIELD')(209,214)
